+++ bayesianbandits/api.py
@@ -71,8 +71,13 @@ import numpy as np
 from numpy.typing import NDArray
 from typing_extensions import Self
 
-from ._arm import Arm, ContextType, TokenType, batch_sample_arms, Learner
+from ._arm import Arm, ContextType, Learner, TokenType
 from ._arm_featurizer import ArmFeaturizer
+from .policies import (  # noqa: F401
+    EpsilonGreedy,
+    ThompsonSampling,
+    UpperConfidenceBound,
+)
 
 
 class PolicyProtocol(Protocol[ContextType, TokenType]):
@@ -1063,385 +1068,3 @@ class LipschitzContextualAgent(Generic[ContextType, TokenType]):
 
         # Decay the shared learner once
         self.learner.decay(X_enriched, decay_rate=decay_rate)
-
-
-class PolicyDefaultUpdate(Generic[ContextType, TokenType]):
-    def update(
-        self,
-        arm: Arm[ContextType, TokenType],
-        X: ContextType,
-        y: NDArray[np.float64],
-        all_arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        sample_weight: Optional[NDArray[np.float64]] = None,
-    ) -> None:
-        """Default update implementation that simply updates the arm."""
-        arm.update(X, y, sample_weight=sample_weight)
-
-
-class EpsilonGreedy(PolicyDefaultUpdate[ContextType, TokenType]):
-    """
-    Policy object for epsilon-greedy.
-
-    Epsilon-greedy chooses the best arm with probability 1 - epsilon and a random
-    arm with probability epsilon.
-
-    Parameters
-    ----------
-    epsilon : float, default=0.1
-        Probability of exploration.
-    samples : int, default=1000
-        Number of samples to use for computing the arm means.
-
-    Notes
-    -----
-    The implementation here is based on the implementation in [1]_.
-
-    References
-    ----------
-    .. [1] Chapelle, Olivier, and Lihong Li. "An empirical evaluation of
-       thompson sampling." Advances in neural information processing systems
-       24 (2011): 2249-2257.
-    """
-
-    def __repr__(self) -> str:
-        return f"EpsilonGreedy(epsilon={self.epsilon}, samples={self.samples})"
-
-    def __init__(self, epsilon: float = 0.1, samples: int = 1000):
-        self.epsilon = epsilon
-        self.samples = samples
-
-    @property
-    def samples_needed(self) -> int:
-        """Number of samples per arm per context needed for decision making."""
-        return self.samples
-
-    @overload
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: None = None,
-    ) -> List[Arm[ContextType, TokenType]]: ...
-
-    @overload
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: int,
-    ) -> List[List[Arm[ContextType, TokenType]]]: ...
-
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: Optional[int] = None,
-    ) -> Union[
-        List[Arm[ContextType, TokenType]], List[List[Arm[ContextType, TokenType]]]
-    ]:
-        """Select arms based on pre-generated samples using epsilon-greedy."""
-        # Compute means across samples dimension
-        means = samples.mean(axis=-1)  # Shape: (n_arms, n_contexts)
-
-        # Apply epsilon-greedy exploration per context
-        # Decide which contexts explore
-        choice_idx_to_explore = rng.random(size=means.shape[1]) < self.epsilon
-
-        # How many arms to mark as infinite
-        k = 1 if top_k is None else top_k
-
-        # For each exploring context, set k random arms to np.inf
-        values = means.copy()
-        for idx, explore in enumerate(choice_idx_to_explore):
-            if explore:
-                random_arms = rng.choice(
-                    means.shape[0],
-                    size=min(k, means.shape[0]),
-                    replace=False,
-                )
-                values[random_arms, idx] = np.inf
-
-        if top_k is None:
-            best_indices = values.argmax(axis=0)
-            return [arms[idx] for idx in best_indices]
-        else:
-            # Use argpartition for efficiency, but handle case where top_k >= n_arms
-            return [
-                [
-                    arms[idx]
-                    for idx in np.argpartition(
-                        -values[:, i], min(top_k, len(arms) - 1)
-                    )[:top_k]
-                ]
-                for i in range(values.shape[1])
-            ]
-
-    @overload
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: None = None,
-    ) -> List[Arm[ContextType, TokenType]]: ...
-
-    @overload
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: int,
-    ) -> List[List[Arm[ContextType, TokenType]]]: ...
-
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: Optional[int] = None,
-    ) -> Union[
-        List[Arm[ContextType, TokenType]], List[List[Arm[ContextType, TokenType]]]
-    ]:
-        """Choose arm(s) using epsilon-greedy."""
-        samples = batch_sample_arms(arms, X, size=self.samples_needed)
-        if samples is None:
-            samples = np.array([arm.sample(X, self.samples_needed) for arm in arms])
-            # Convert from (n_arms, size, n_contexts) to (n_arms, n_contexts, size)
-            samples = samples.transpose(0, 2, 1)
-        return self.select(samples, arms, rng, top_k)
-
-
-class ThompsonSampling(PolicyDefaultUpdate[ContextType, TokenType]):
-    """
-    Policy object for Thompson sampling.
-
-    Thompson sampling chooses the best arm with probability equal to the
-    probability that the arm is the best arm. That is, it takes a sample from
-    each arm's posterior distribution and chooses the arm with the highest
-    sample.
-
-    Notes
-    -----
-    The implementation here is based on the implementation in [1]_.
-
-    References
-    ----------
-    .. [1] Chapelle, Olivier, and Lihong Li. "An empirical evaluation of
-       thompson sampling." Advances in neural information processing systems
-       24 (2011): 2249-2257.
-    """
-
-    def __repr__(self) -> str:
-        return "ThompsonSampling()"
-
-    @property
-    def samples_needed(self) -> int:
-        """Number of samples per arm per context needed for decision making."""
-        return 1
-
-    @overload
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: None = None,
-    ) -> List[Arm[ContextType, TokenType]]: ...
-
-    @overload
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: int,
-    ) -> List[List[Arm[ContextType, TokenType]]]: ...
-
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: Optional[int] = None,
-    ) -> Union[
-        List[Arm[ContextType, TokenType]], List[List[Arm[ContextType, TokenType]]]
-    ]:
-        """Select arms based on pre-generated samples."""
-        # samples shape: (n_arms, n_contexts, 1)
-        values = samples[..., 0]  # Shape: (n_arms, n_contexts)
-
-        if top_k is None:
-            best_indices = values.argmax(axis=0)
-            return [arms[idx] for idx in best_indices]
-        else:
-            # Use argpartition for efficiency, but handle case where top_k >= n_arms
-            return [
-                [
-                    arms[idx]
-                    for idx in np.argpartition(
-                        -values[:, i], min(top_k, len(arms) - 1)
-                    )[:top_k]
-                ]
-                for i in range(values.shape[1])
-            ]
-
-    @overload
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: None = None,
-    ) -> List[Arm[ContextType, TokenType]]: ...
-
-    @overload
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: int,
-    ) -> List[List[Arm[ContextType, TokenType]]]: ...
-
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: Optional[int] = None,
-    ) -> Union[
-        List[Arm[ContextType, TokenType]], List[List[Arm[ContextType, TokenType]]]
-    ]:
-        """Choose arm(s) using Thompson sampling."""
-        samples = batch_sample_arms(arms, X, size=self.samples_needed)
-        if samples is None:
-            samples = np.array([arm.sample(X, self.samples_needed) for arm in arms])
-            # Convert from (n_arms, size, n_contexts) to (n_arms, n_contexts, size)
-            samples = samples.transpose(0, 2, 1)
-        return self.select(samples, arms, rng, top_k)
-
-
-class UpperConfidenceBound(PolicyDefaultUpdate[ContextType, TokenType]):
-    """
-    Policy object for upper confidence bound.
-
-    Upper confidence bound takes `samples` samples from each arm's posterior
-    distribution and chooses the arm with the highest upper bound, as defined
-    by the `alpha` parameter.
-
-    Parameters
-    ----------
-    alpha : float, default=0.68
-        Confidence level (one-sided)
-    samples : int, default=1000
-        Number of samples to use for computing the arm upper bounds.
-
-    Notes
-    -----
-    The implementation here is based on the implementation in [1]_.
-
-    References
-    ----------
-    .. [1] Chapelle, Olivier, and Lihong Li. "An empirical evaluation of
-       thompson sampling." Advances in neural information processing systems
-       24 (2011): 2249-2257.
-    """
-
-    def __repr__(self) -> str:
-        return f"UpperConfidenceBound(alpha={self.alpha}, samples={self.samples})"
-
-    def __init__(self, alpha: float = 0.68, samples: int = 1000):
-        self.alpha = alpha
-        self.samples = samples
-
-    @property
-    def samples_needed(self) -> int:
-        """Number of samples per arm per context needed for decision making."""
-        return self.samples
-
-    @overload
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: None = None,
-    ) -> List[Arm[ContextType, TokenType]]: ...
-
-    @overload
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: int,
-    ) -> List[List[Arm[ContextType, TokenType]]]: ...
-
-    def select(
-        self,
-        samples: NDArray[np.float64],  # Shape: (n_arms, n_contexts, samples_needed)
-        arms: List[Arm[ContextType, TokenType]],
-        rng: np.random.Generator,
-        top_k: Optional[int] = None,
-    ) -> Union[
-        List[Arm[ContextType, TokenType]], List[List[Arm[ContextType, TokenType]]]
-    ]:
-        """Select arms based on pre-generated samples using upper confidence bounds."""
-        # Compute upper confidence bounds using quantiles
-        # samples shape: (n_arms, n_contexts, samples_needed)
-        ucb_values = np.quantile(samples, self.alpha, axis=-1)
-
-        if top_k is None:
-            best_indices = ucb_values.argmax(axis=0)
-            return [arms[idx] for idx in best_indices]
-        else:
-            # Use argpartition for efficiency, but handle case where top_k >= n_arms
-            return [
-                [
-                    arms[idx]
-                    for idx in np.argpartition(
-                        -ucb_values[:, i], min(top_k, len(arms) - 1)
-                    )[:top_k]
-                ]
-                for i in range(ucb_values.shape[1])
-            ]
-
-    @overload
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: None = None,
-    ) -> List[Arm[ContextType, TokenType]]: ...
-
-    @overload
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: int,
-    ) -> List[List[Arm[ContextType, TokenType]]]: ...
-
-    def __call__(
-        self,
-        arms: List[Arm[ContextType, TokenType]],
-        X: ContextType,
-        rng: np.random.Generator,
-        top_k: Optional[int] = None,
-    ) -> Union[
-        List[Arm[ContextType, TokenType]], List[List[Arm[ContextType, TokenType]]]
-    ]:
-        """Choose arm(s) using upper confidence bound."""
-        samples = batch_sample_arms(arms, X, size=self.samples_needed)
-        if samples is None:
-            samples = np.array([arm.sample(X, self.samples_needed) for arm in arms])
-            # Convert from (n_arms, size, n_contexts) to (n_arms, n_contexts, size)
-            samples = samples.transpose(0, 2, 1)
-        return self.select(samples, arms, rng, top_k)
