+++ bayesianbandits/policies/__init__.py
@@ -1,5 +1,11 @@
 from ._exp3a import EXP3A
+from ._epsilon_greedy import EpsilonGreedy
+from ._thompson_sampling import ThompsonSampling
+from ._upper_confidence_bound import UpperConfidenceBound
 
 __all__ = [
     "EXP3A",
+    "EpsilonGreedy", 
+    "ThompsonSampling",
+    "UpperConfidenceBound",
 ]
